requirements.txt
discord.py==2.4.0
groq==0.9.0
python-dotenv==1.0.1